dsp: asm: Avoid log spam for validation of payload size
On three occasions, the check of payload_size is combined with other checks. This results in errors for payload size being too small, when it is actually not the case.

Separate the check for paylod_size from other checks to avoid spamming the log for no (or at least not the right) reason.

Also added another instance of error reporting if payload_size is too small as it was missing.